Moves rules and ranges to sub directories
