+++ src/test/java/pro/trevor/tankgame/rule/CouncilGrantLifeTest.java
@@ -12,7 +12,7 @@ import pro.trevor.tankgame.rule.impl.version3.Tank;
 import pro.trevor.tankgame.state.meta.Council;
 
 import static org.junit.jupiter.api.Assertions.*;
-import static pro.trevor.tankgame.rule.impl.shared.PlayerRules.GetRuleCofferCostGrantLife;
+import static pro.trevor.tankgame.rule.impl.shared.rule.PlayerRules.GetRuleCofferCostGrantLife;
 
 public class CouncilGrantLifeTest {
